Move AMR* to model
